# lib/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Agendamento v2",
    "version": "1.0.0",
    "description": "API REST completa para sistema de agendamentos com autenticação via API Key",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key no formato: sk_<apiClientId>_<random>"
      },
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token do Supabase (apenas para endpoints administrativos)"
      }
    },
    "schemas": {
...
      "Professional": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/professionals/{id}": {
      "get": {
        "tags": ["Agendamentos"],
        "summary": "Buscar Professional por ID",
        "description": "Busca um professional específico por ID na company associada à API Key.\n\n**Exemplo de requisição:**\n```bash\ncurl -X GET http://localhost:3000/api/v1/professionals/123e4567-e89b-12d3-a456-426614174000 \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\"\n```",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Professional encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/Professional"}
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Dr. João Silva",
                    "email": "joao@example.com",
                    "phone": "+5511999999999",
                    "created_at": "2024-01-01T00:00:00Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "404": {
            "description": "Professional não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Professional not found"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Agendamentos"],
        "summary": "Atualizar Professional",
        "description": "Atualiza um professional existente na company associada à API Key.\n\n**Exemplo de requisição:**\n```bash\ncurl -X PUT http://localhost:3000/api/v1/professionals/123e4567-e89b-12d3-a456-426614174000 \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"name\": \"Dr. João Silva\",\n    \"email\": \"joao.updated@example.com\",\n    \"phone\": \"+5511888888888\"\n  }'\n```",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Dr. João Silva"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "nullable": true,
                    "example": "joao@example.com"
                  },
                  "phone": {
                    "type": "string",
                    "nullable": true,
                    "example": "+5511999999999"
                  }
                }
              },
              "example": {
                "name": "Dr. João Silva",
                "email": "joao.updated@example.com",
                "phone": "+5511888888888"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Professional atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"$ref": "#/components/schemas/Professional"}
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Dr. João Silva",
                    "email": "joao.updated@example.com",
                    "phone": "+5511888888888",
                    "created_at": "2024-01-01T00:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Validation error",
                  "errors": {
                    "email": ["Invalid email"]
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "404": {
            "description": "Professional não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Professional not found"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Agendamentos"],
        "summary": "Deletar Professional",
        "description": "Deleta um professional da company associada à API Key.\n\n**Exemplo de requisição:**\n```bash\ncurl -X DELETE http://localhost:3000/api/v1/professionals/123e4567-e89b-12d3-a456-426614174000 \\\n  -H \"Authorization: Bearer <API_KEY>\" \\\n  -H \"Content-Type: application/json\"\n```",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "200": {
            "description": "Professional deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {"type": "string", "example": "Professional deleted successfully"}
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "data": {
                    "message": "Professional deleted successfully"
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key inválida ou ausente",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Unauthorized: Invalid or missing API key"
                }
              }
            }
          },
          "404": {
            "description": "Professional não encontrado",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Professional not found"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}